Rename formula_support to keg_only reason, refactor slightly
