--- Library/Homebrew/extend/os/formula_support.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require "extend/os/mac/formula_support" if OS.mac?
