+++ Library/Homebrew/extend/os/keg_only_reason.rb
@@ -1,0 +1,4 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "extend/os/mac/keg_only_reason" if OS.mac?
