+++ Library/Homebrew/extend/os/mac/keg_only_reason.rb
